+++ xarray/core/dataset.py
@@ -13,6 +13,7 @@ import pandas as pd
 from . import ops
 from . import utils
 from . import groupby
+from . import rolling
 from . import indexing
 from . import alignment
 from . import formatting
@@ -307,6 +308,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
     coordinates used for label based indexing.
     """
     groupby_cls = groupby.DatasetGroupBy
+    rolling_cls = rolling.DatasetRolling
 
     def __init__(self, data_vars=None, coords=None, attrs=None,
                  compat='broadcast_equals'):